python/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
astropy==6.1.1
xisf
numpy
Pillow
tifffile
imagecodecs
exifread
rawpy
aiofiles~=24.1
aiohttp>=3.10.9
pyephem>=9.99
pytz>=2024.2
setuptools>=75.1
typeguard>=4.3
openmeteo-requests~=1.3
pandas>=2.1
tabulate~=0.9
pyastroweatherio==0.74.0
python-dateutil